requirements.txt
Scrapy>=2.0.0
SQLAlchemy>=1.3.13
python_dateutil>=2.8.1
ipython
ipykernel

